Moves the committee member service to its own service file
